Add base test class
