+++ usr/lib/python3/dist-packages/stdisplay/tests/stcat.py
@@ -1,67 +1,31 @@
 #!/usr/bin/env python3
 
-import os
-import shutil
-import sys
-import tempfile
 import unittest
-from unittest.mock import patch
-from test.support import captured_stdout, captured_stdin
+import stdisplay.tests
 
-
-class TestSTCat(unittest.TestCase):
+class TestSTCat(stdisplay.tests.TestST):
     """
     Test stcat.
     """
 
     def setUp(self):
-        self.temp_files = []
-        contents = ["a b\n", "c d"]
-        i = 0
-        self.temp_dir = tempfile.mkdtemp(prefix="test_stcat_")
-        while i < 3:
-            self.temp_files.append(os.path.join(self.temp_dir, str(i)))
-            with open(self.temp_files[i], "w", encoding="utf-8") as file:
-                if i == 0:
-                    file.write("")
-                elif i == 1:
-                    file.write("".join(contents))
-                elif i == 2:
-                    file.write("".join(contents) + "\n")
-                file.flush()
-                file.close()
-            i += 1
+        self.module = "stcat"
+        super().setUp()
 
-    def tearDown(self):
-        del sys.modules["stdisplay.stcat"]
-        shutil.rmtree(self.temp_dir)
-
-    def _del_module(self):
-        module_name = "stdisplay.stcat"
-        if module_name in sys.modules:
-            del sys.modules[module_name]
-        globals().pop(module_name, None)
-
-    def _test_stcat(self, argv=None, incoming=None):
+    def test_stcat_stdin(self):
         """
-        Helper function to pass stdin.
+        Test passing stdin.
         """
-        self._del_module()
-        if argv is None:
-            argv = ["stcat.py"]
-        else:
-            argv = ["stcat.py"] + argv
-        with patch.object(
-            sys, "argv", argv
-        ), captured_stdout() as stdout, captured_stdin() as stdin:
-            import stdisplay.stcat  # pylint: disable=import-outside-toplevel
-
-            if incoming is not None:
-                stdin.write(incoming)
-            stdin.seek(0)
-            stdisplay.stcat.main()
-        result = stdout.getvalue()  # pylint: disable=no-member
-        return result
+        self.assertEqual("", self._test_util())
+        self.assertEqual("", self._test_util(incoming=""))
+        self.assertEqual("", self._test_util(argv=["-"]))
+        self.assertEqual(
+            "Hello world", self._test_util(incoming="Hello world")
+        )
+        self.assertEqual(
+            "Hello world",
+            self._test_util(argv=["-"], incoming="Hello world"),
+        )
 
     def test_stcat_file(self):
         """
@@ -75,22 +39,7 @@ class TestSTCat(unittest.TestCase):
         ]
         for text, argv in cases:
             with self.subTest(text=text, argv=argv):
-                self.assertEqual(text, self._test_stcat(argv=argv))
-
-    def test_stcat_stdin(self):
-        """
-        Test passing stdin.
-        """
-        self.assertEqual("", self._test_stcat())
-        self.assertEqual("", self._test_stcat(incoming=""))
-        self.assertEqual("", self._test_stcat(argv=["-"]))
-        self.assertEqual(
-            "Hello world", self._test_stcat(incoming="Hello world")
-        )
-        self.assertEqual(
-            "Hello world",
-            self._test_stcat(argv=["-"], incoming="Hello world"),
-        )
+                self.assertEqual(text, self._test_util(argv=argv))
 
 
 if __name__ == "__main__":
